Base/FPGA: second ethernet again
Prior bitstream did not detect device. For some reason
it needs ETH/MDIO devices enabled for network LEDs to
come on and link status to be detected.

However even with this IP connectivity is not working.

Also confirmed that dropping this bitstream into the
krtkl image it does not work, so the bitstream still
may issues. However, this could also reflect a change
in IP/drivers since then.