U-boot: check in SnickerdoodleBlack defconfig
Have to apply this as a patch to the meta layer. This copies the
approach in the official Pynq boards. That being said, it does not
seem like these changes are actually applied. The device tree
seems to come from the system-user.dtsi file and the defconfig
settings are also not used.